requirements.txt
py==1.6.0
pytest==3.8.0
pytest-mock==1.10.0
wheel==0.32.1
paho-mqtt==1.4.0
transitions==0.6.8
enum34==1.1.6
flake8
msrest
six
mock
black; python_version > '3.6'